extract tests to separate module
